id modificados

+++ Model/Admin/crear/rango.php
@@ -5,16 +5,27 @@ $DataBase = new Database;
 $con = $DataBase->conectar();
 
 if ($_SERVER["REQUEST_METHOD"] == "POST") {
+    $id_rango = $_POST["id_rango"]; // Recibe el id_rango ingresado manualmente
     $nombre = $_POST["nombre"];
     $foto = file_get_contents($_FILES['foto']['tmp_name']);
 
-    $sql = "INSERT INTO rango (nombre, foto) VALUES (:nombre, :foto)";
+    // Verificar si el id_rango ya existe
+    $sql_check_id = "SELECT COUNT(*) FROM rango WHERE id_rango = :id_rango";
+    $stmt_check_id = $con->prepare($sql_check_id);
+    $stmt_check_id->execute(array(':id_rango' => $id_rango));
+    $id_exists = $stmt_check_id->fetchColumn();
+
+    if ($id_exists) {
+        echo "<script>alert('El ID de rango ya existe. Por favor, elige otro.'); window.location='../visualizar/rango.php';</script>";
+        exit();
+    }
+
+    $sql = "INSERT INTO rango (id_rango, nombre, foto) VALUES (:id_rango, :nombre, :foto)";
     $stmt = $con->prepare($sql);
 
-    $stmt->execute(array(':nombre' => $nombre, ':foto' => $foto));
+    $stmt->execute(array(':id_rango' => $id_rango, ':nombre' => $nombre, ':foto' => $foto));
 
     // Redireccionar a la página actual para actualizar la tabla
-
     echo "<script>alert('Rango creado.'); window.location='../visualizar/rango.php';</script>";
     exit();
 }
@@ -24,6 +35,10 @@ if ($_SERVER["REQUEST_METHOD"] == "POST") {
         <div class="col-md-6">
             <h2 class="text-center">Crear Rango</h2>
             <form method="post" enctype="multipart/form-data">
+                <div class="form-group">
+                    <label for="id_rango">ID Rango:</label>
+                    <input type="number" class="form-control" name="id_rango" required>
+                </div>
                 <div class="form-group">
                     <label for="nombre">Nombre:</label>
                     <input type="text" class="form-control" name="nombre" required>
@@ -39,6 +54,5 @@ if ($_SERVER["REQUEST_METHOD"] == "POST") {
         </div>
     </div>
 </div>
-
 
 <?php include "../template/footer.php"; ?>
